made Circle.id optionally custom

+++ server/api/models/circle.py
@@ -1,9 +1,9 @@
-from django.db.models import Model, CharField, DateField, DateTimeField, SmallIntegerField
+from django.db.models import Model, AutoField, CharField, DateField, DateTimeField
 from .budget import Budget
 
 
 class Circle(Model):
-    id = SmallIntegerField(primary_key=True)
+    id = AutoField(primary_key=True)
     manager = CharField(max_length=256, null=True)
     title = CharField(max_length=512, null=False)
     funding_source = CharField(max_length=512, null=True, default='')
